% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.instruments.R
\docType{data}
\name{currencies}
\alias{currencies}
\title{currency metadata to be used by \code{\link{load.instruments}}}
\description{
currency metadata to be used by \code{\link{load.instruments}}
}
\keyword{data}
