\name{getInstrument}
\alias{getInstrument}
\title{primare accessor function for getting objects of type 'instrument'...}
\usage{getInstrument(x, Dates)}
\description{primare accessor function for getting objects of type 'instrument'}
\arguments{\item{x}{string identifier of instrument to retrieve}
\item{Dates}{date range to retrieve 'as of', may not currently be implemented}}
