% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument.R
\name{is.currency.name}
\alias{is.currency.name}
\title{check each element of a character vector to see if it is either the
primary_id or an identifier of a \code{\link{currency}}}
\usage{
is.currency.name(x)
}
\arguments{
\item{x}{character vector}
}
\value{
A logical vector indicating whether each element of `x` identifies
  a defined currency instrument. Non-character input returns `FALSE`.
}
\description{
check each element of a character vector to see if it is either the
primary_id or an identifier of a \code{\link{currency}}
}
