% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument.R
\name{is.currency}
\alias{is.currency}
\title{class test for object supposedly of type 'currency'}
\usage{
is.currency(x)
}
\arguments{
\item{x}{object to test for type}
}
\description{
class test for object supposedly of type 'currency'
}
