\name{is.instrument.name}
\alias{is.instrument.name}
\title{check each element of a character vector to see if it is either the
primary_id or an identifier of an \code{\link{instrument}}}
\usage{
  is.instrument.name(x)
}
\arguments{
  \item{x}{character vector}
}
\value{
  logical vector
}
\description{
  check each element of a character vector to see if it is
  either the primary_id or an identifier of an
  \code{\link{instrument}}
}

