% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.instrument}
\alias{is.instrument}
\title{class test for object supposedly of type 'instrument'}
\usage{
is.instrument(x)
}
\arguments{
\item{x}{object to test for type}
}
\description{
class test for object supposedly of type 'instrument'
}

