% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.instruments.R
\docType{data}
\name{root_contracts}
\alias{root_contracts}
\title{future metadata to be used by \code{\link{load.instruments}}}
\description{
future metadata to be used by \code{\link{load.instruments}}
}
\keyword{data}
