\name{exchange_rate}
\alias{exchange_rate}
\title{constructor for spot exchange rate instruments...}
\usage{exchange_rate(primary_id, currency, second_currency, identifiers, ...)}
\description{constructor for spot exchange rate instruments}
\arguments{\item{primary_id}{string identifier, usually expressed as a currency pair 'USDYEN' or 'EURGBP'}
\item{currency}{string identifying front currency}
\item{second_currency}{string identifying second currency}
\item{identifiers}{character vector of any other identifiers that should also be stored for this instrument}
\item{...}{any other passthru parameters}}
