\name{getInstrument}
\alias{getInstrument}
\title{primary accessor function for getting objects of type 'instrument'...}
\usage{getInstrument(x, Dates, silent=FALSE)}
\description{primary accessor function for getting objects of type 'instrument'}
\arguments{\item{x}{string identifier of instrument to retrieve}
\item{Dates}{date range to retrieve 'as of', may not currently be implemented}
\item{silent}{if TRUE, will not warn on failure, default FALSE}}
