\name{is.currency.name}
\alias{is.currency.name}
\title{check each element of a character vector to see if it is either the
primary_id or an identifier of a \code{\link{currency}}}
\usage{
  is.currency.name(x)
}
\arguments{
  \item{x}{character vector}
}
\description{
  check each element of a character vector to see if it is
  either the primary_id or an identifier of a
  \code{\link{currency}}
}

