% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_by_currency.R
\name{ls_by_currency}
\alias{ls_by_currency}
\alias{rm_by_currency}
\alias{ls_USD}
\alias{ls_AUD}
\alias{ls_GBP}
\alias{ls_CAD}
\alias{ls_EUR}
\alias{ls_JPY}
\alias{ls_CHF}
\alias{ls_HKD}
\alias{ls_SEK}
\alias{ls_NZD}
\title{shows or removes instruments of given currency denomination(s)}
\usage{
ls_by_currency(currency, pattern = NULL, match = TRUE, show.currencies = FALSE)

rm_by_currency(x, currency, keep.currencies = TRUE)

ls_USD(pattern = NULL, match = TRUE, show.currencies = FALSE)

ls_AUD(pattern = NULL, match = TRUE, show.currencies = FALSE)

ls_GBP(pattern = NULL, match = TRUE, show.currencies = FALSE)

ls_CAD(pattern = NULL, match = TRUE, show.currencies = FALSE)

ls_EUR(pattern = NULL, match = TRUE, show.currencies = FALSE)

ls_JPY(pattern = NULL, match = TRUE, show.currencies = FALSE)

ls_CHF(pattern = NULL, match = TRUE, show.currencies = FALSE)

ls_HKD(pattern = NULL, match = TRUE, show.currencies = FALSE)

ls_SEK(pattern = NULL, match = TRUE, show.currencies = FALSE)

ls_NZD(pattern = NULL, match = TRUE, show.currencies = FALSE)
}
\arguments{
\item{currency}{chr vector of names of currency}

\item{pattern}{an optional regular expression.  Only names matching
\sQuote{pattern} are returned.}

\item{match}{exact match?}

\item{show.currencies}{include names of currency instruments in the returned
names?}

\item{x}{what to remove. chr vector.}

\item{keep.currencies}{Do not delete currency instruments when deleting
multiple instruments.}
}
\value{
ls_ functions return vector of instrument names rm_ functions return
invisible / called for side-effect.
}
\description{
ls_ functions get names of instruments denominated in a given currency (or
currencies) rm_ functions remove instruments of a given currency
}
\examples{
backup_file <- tempfile(fileext = ".RData")
saveInstruments(backup_file)

rm_instruments(keep.currencies = FALSE)
currency(c("USD", "CAD", "GBP"))
stock(c("CM", "CNQ"), "CAD")
stock(c("BARC", "BET"), "GBP")
stock(c("DIA", "SPY"), "USD")

ls_by_currency(c("CAD", "GBP"))
ls_USD()
ls_CAD()

rm_by_currency(currency = "CAD")

reloadInstruments(backup_file)
unlink(backup_file)

}
\seealso{
ls_instruments, ls_currencies, rm_instruments, rm_currencies,
twsInstrument, instrument
}
\author{
Garrett See
}
