% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument.R
\name{sort.instrument}
\alias{sort.instrument}
\title{instrument class sort method}
\usage{
\method{sort}{instrument}(x, decreasing = FALSE, na.last = NA, ...)
}
\description{
instrument class sort method
}
\author{
Garrett See
}
\keyword{internal}
