% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sort.instrument}
\alias{sort.instrument}
\title{instrument class sort method}
\usage{
\method{sort}{instrument}(x, decreasing = FALSE, na.last = NA, ...)
}
\description{
instrument class sort method
}
\author{
Garrett See
}
\keyword{internal}

