% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volep.R
\name{volep}
\alias{volep}
\title{generate endpoints for volume bars}
\usage{
volep(x, units)
}
\arguments{
\item{x}{time series containing 'Volume' column}

\item{units}{volume sum to mark for bars}
}
\description{
generate endpoints for volume bars
}
\author{
Joshua Ulrich
}
